Improve support for clang-cl (#186)
* Improve build configuration for Clang on Windows

* Add a launch config for recon.exe that includes conversion

* Fix various warnings that occur in clang-cl

* Clean up and support windows msvc and clang and ubuntu(wsl) with gcc

* Improve test failure behavior for runtime FS tests

+++ source/bin_shell/private/editors/filetree_editor.cpp
@@ -61,7 +61,7 @@ void up::shell::FileTreeEditor::content() {
             ImGui::Text("-");
         }
         else {
-            ImGui::Text("%lu", cached.size);
+            ImGui::Text("%lu", narrow_cast<unsigned long>(cached.size));
         }
         ImGui::NextColumn();
         auto ext = path::extension(cached.name);
